+++ src/game/research/TechCostCalculator.cpp
@@ -16,18 +16,18 @@ TechCostCalculator::TechCostCalculator(const TechCostConfig_t& rConfig, LuaRunti
 
 int TechCostCalculator::CalculateCost(const TechConfig_t& rTech, const TechCostInputs_t& rInputs) const
 {
-    std::unordered_map<std::string, int> vars = {
-        {"techs",               rInputs.techs},
-        {"most_techs",          rInputs.mostTechs},
-        {"diff",                rInputs.diff},
-        {"turns",               rInputs.turns},
-        {"is_ai",               rInputs.bIsAI ? 1 : 0},
-        {"tech_stagnation",     rInputs.bTechStagnation ? 1 : 0},
-        {"research_modifier",   rInputs.researchModifier},
-        {"world_size_modifier", rInputs.worldSizeModifier},
-        {"faction_modifier",    rInputs.factionTechCostModifier},
-        {"alphax_modifier",     rInputs.alphaxTechCostModifier},
-        {"base_cost",           rTech.cost},
+    std::unordered_map<std::string, double> vars = {
+        {"techs",               static_cast<double>(rInputs.techs)},
+        {"most_techs",          static_cast<double>(rInputs.mostTechs)},
+        {"diff",                static_cast<double>(rInputs.diff)},
+        {"turns",               static_cast<double>(rInputs.turns)},
+        {"is_ai",               rInputs.bIsAI ? 1.0 : 0.0},
+        {"tech_stagnation",     rInputs.bTechStagnation ? 1.0 : 0.0},
+        {"research_modifier",   static_cast<double>(rInputs.researchModifier)},
+        {"world_size_modifier", static_cast<double>(rInputs.worldSizeModifier)},
+        {"faction_modifier",    static_cast<double>(rInputs.factionTechCostModifier)},
+        {"alphax_modifier",     static_cast<double>(rInputs.alphaxTechCostModifier)},
+        {"base_cost",           static_cast<double>(rTech.cost)},
     };
 
     const int cost = m_pLua->EvalInt(m_pConfig->costFormula, vars);
